@@ -15,10 +15,10 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 1.0.0+1
+version: 1.1.0+2
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -27,17 +27,20 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  pdf: ^2.1.0
-  cupertino_icons: ^1.0.0
-  provider: ^4.3.3
-  filesize: ^1.0.4
-  path_provider:
-  flutter_share: ^1.0.5
-  random_string: ^2.1.0
-  flutter_archive: ^2.0.2
-  string_validator: ^0.1.4
-  multi_image_picker: ^4.7.15
-  flutter_native_image: ^0.0.5
+  pdf: ^3.4.2
+  cupertino_icons: ^1.0.3
+  provider: ^5.0.0
+  filesize: ^2.0.1
+  path_provider: ^2.0.2
+  share: ^2.0.4
+  random_string: ^2.2.0-nullsafety
+  flutter_archive: ^4.0.1
+  string_validator: ^0.3.0
+  multi_image_picker: ^4.8.01
+  flutter_native_image: ^0.0.6+1
+  tutorial_coach_mark: ^1.0.3
+  shared_preferences: ^2.0.6
+
 #  downloads_path_provider: ^0.1.0
 
 
